requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt
django-cors-headers
django-filter
drf-spectacular
celery[redis]>=5.3
django-celery-beat
redis
psycopg2-binary
cryptography
factory-boy
faker
pytest
pytest-django
pytest-cov
locust
bandit
